Move tests to individual files to make it easier to run tests in isolation
